requirements.txt
streamlit
sentence-transformers
chromadb
faiss-cpu
llama-index
langchain
openai
torch
transformers
